Move replaceChunkMap to other mixin
